🔧 Refactor: Divide logic for frontend and backend • Define logic for both • Define hooks and ui components for frontend • Define api and logic for backend • Fix mail in package.json
