+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "client",
-  "author": "Daniel Bakas <daniel@semanytk.com>",
+  "author": "Daniel Bakas <daniel@semantyk.com>",
   "bugs": "https://github.com/semantyk/client/issues",
   "description": "Client for semantyk.com | Semantyk",
   "license": "CC-BY-NC-SA-4.0",
   "maintainers": [
-    "Daniel Bakas <daniel@semanytk.com>"
+    "Daniel Bakas <daniel@semantyk.com>"
   ],
   "private": true,
   "repository": {
@@ -22,7 +22,8 @@
     "react": "^18.2.0",
     "react-bootstrap": "^2.9.1",
     "react-dom": "^18.2.0",
-    "recoil": "^0.7.7"
+    "recoil": "^0.7.7",
+    "sharp": "^0.33.0"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.1.5",
@@ -30,7 +31,7 @@
     "@testing-library/user-event": "^14.5.1",
     "eslint": "^8.55.0",
     "eslint-config-next": "14.0.3",
-    "next": "^14.0.3",
+    "next": "^14.0.4",
     "vitest": "^1.0.4"
   },
   "scripts": {
