+++ src/app/not-found.jsx
@@ -1,13 +1,13 @@
 /*
  * –––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
- * # `not-found.jsx`
+ * # `not-found.jsx` | `app`
  * client | Semantyk
  *
  * Created: Dec 06, 2023
  * Modified: Dec 06, 2023
  *
  * Author(s): Semantyk Team
- * Maintainer(s):
+ * Maintainer(s): Daniel Bakas <https://id.danielbakas.com
  *
  * Copyright © Semantyk 2023. All rights reserved.
  * –––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
