+++ src/app/page.css
@@ -1,10 +1,10 @@
 /*
  * –––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
- * # `page.css` | `data`
+ * # `page.css` | 'app'
  * client | Semantyk
  *
  * Created: Nov 30, 2023
- * Modified: Nov 30, 2023
+ * Modified: Dec 10, 2023
  *
  * Author(s): Semantyk Team
  * Maintainer(s): Daniel Bakas <https://id.danielbakas.com>
