+++ src/app/page.jsx
@@ -1,10 +1,10 @@
 /*
  * –––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
- * # `page.jsx` | `data`
+ * # `page.jsx` | `app`
  * client | Semantyk
  *
  * Created: Nov 30, 2023
- * Modified: Dic 6, 2023
+ * Modified: Dec 10, 2023
  *
  * Author(s): Semantyk Team
  * Maintainer(s): Daniel Bakas <https://id.danielbakas.com>
@@ -17,8 +17,7 @@
 import React from "react";
 //* Local Imports
 import "./page.css";
-import Footer from "@/components/molecules/footer/component";
-import AuthButton from "@/components/atoms/authButton";
+import Footer from "@/frontend/ui/components/molecules/footer/component";
 
 
 //* Main
@@ -27,7 +26,6 @@ export default function Page({ children }) {
     return (
         <div id="Page">
             {children}
-            <AuthButton/>
             <Footer/>
         </div>
     );
