+++ src/backend/api/knowledge/getAppKnowledge.js
@@ -1,9 +1,9 @@
 /*
  * –––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
- * # `layout.jsx` | `app`
+ * # `getAppKnowledge.js` | `knowledge`
  * client | Semantyk
  *
- * Created: Nov 30, 2023
+ * Created: Dec 10, 2023
  * Modified: Dec 10, 2023
  *
  * Author(s): Semantyk Team
@@ -14,25 +14,18 @@
  */
 
 //* Imports
-import "bootstrap/dist/css/bootstrap.min.css";
-import React from "react";
-//* Local Imports
-import { getMetadata } from "@/backend/logic/seo";
-import StateManager from "@/frontend/logic/state";
-
+import { APP_WEBID, APP_WEBID_DOC } from "@/backend/logic/kgm/nodes";
+import { getThing } from "@/backend/api/solid/getThing";
+import { getProperties } from "@/backend/api/solid/getProperties";
+import { appShape } from "@/backend/logic/kgm/shapes";
 
 //* Main
-export async function generateMetadata() {return await getMetadata();}
-
-export default function RootLayout({ children }) {
+export async function getAppKnowledge(fetch) {
+    // Logic
+    const document = APP_WEBID_DOC;
+    const uri = APP_WEBID;
+    const thing = await getThing(fetch, document, uri);
+    const lang = "en";
     // Return
-    return (
-        <html lang="en">
-        <body className={`bg-dark text-light`}>
-        <StateManager>
-            {children}
-        </StateManager>
-        </body>
-        </html>
-    );
+    return getProperties(thing, appShape, lang);
 }
