+++ src/backend/api/solid/getURLAll.js
@@ -1,9 +1,9 @@
 /*
  * –––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
- * # `layout.jsx` | `app`
+ * # `getURLAll.js` | `solid`
  * client | Semantyk
  *
- * Created: Nov 30, 2023
+ * Created: Dec 05, 2023
  * Modified: Dec 10, 2023
  *
  * Author(s): Semantyk Team
@@ -14,25 +14,9 @@
  */
 
 //* Imports
-import "bootstrap/dist/css/bootstrap.min.css";
-import React from "react";
-//* Local Imports
-import { getMetadata } from "@/backend/logic/seo";
-import StateManager from "@/frontend/logic/state";
-
+import { getUrlAll } from "@inrupt/solid-client";
 
 //* Main
-export async function generateMetadata() {return await getMetadata();}
-
-export default function RootLayout({ children }) {
-    // Return
-    return (
-        <html lang="en">
-        <body className={`bg-dark text-light`}>
-        <StateManager>
-            {children}
-        </StateManager>
-        </body>
-        </html>
-    );
-}
+export const getURLAll = (thing, property) => {
+    return getUrlAll(thing, property);
+};
