+++ src/backend/logic/auth.js
@@ -1,25 +1,25 @@
 /*
  * –––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
- * # `service.js` | `auth`
+ * # `auth.js` | `logic`
  * client | Semantyk
  *
  * Created: Dec 04, 2023
- * Modified: Dec 04, 2023
+ * Modified: Dec 10, 2023
  *
  * Author(s): Semantyk Team
- * Maintainer(s):
+ * Maintainer(s): Daniel Bakas <https://id.danielbakas.com>
  *
  * Copyright © Semantyk 2023. All rights reserved.
  * –––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
  */
 
 //* Imports
-import { CLIENT_ID, POD_PROVIDER_URI } from "@/services/app/nodes";
+import { CLIENT_WEBID, POD_PROVIDER_WEBID } from "@/backend/logic/kgm/nodes";
 
 export async function getOptions() {
     // Logic
-    const clientId = CLIENT_ID.value;
-    const oidcIssuer = POD_PROVIDER_URI.value;
+    const clientId = CLIENT_WEBID;
+    const oidcIssuer = POD_PROVIDER_WEBID;
     const redirectUrl = window.location.href;
     // Return
     return {
