+++ src/backend/logic/kgm/nodes.js
@@ -1,9 +1,9 @@
 /*
  * –––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
- * # `layout.jsx` | `app`
+ * # `nodes.js` | `kgm`
  * client | Semantyk
  *
- * Created: Nov 30, 2023
+ * Created: Dec 05, 2023
  * Modified: Dec 10, 2023
  *
  * Author(s): Semantyk Team
@@ -13,26 +13,13 @@
  * –––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
  */
 
-//* Imports
-import "bootstrap/dist/css/bootstrap.min.css";
-import React from "react";
-//* Local Imports
-import { getMetadata } from "@/backend/logic/seo";
-import StateManager from "@/frontend/logic/state";
-
-
 //* Main
-export async function generateMetadata() {return await getMetadata();}
-
-export default function RootLayout({ children }) {
-    // Return
-    return (
-        <html lang="en">
-        <body className={`bg-dark text-light`}>
-        <StateManager>
-            {children}
-        </StateManager>
-        </body>
-        </html>
-    );
-}
+// Namespaces
+export const ARCHIVE = "https://www.semantyk.com/archive/";
+export const SCHEMA = "http://schema.org/";
+// WebIDs
+export const APP_WEBID = "https://id.inrupt.com/semantyk";
+export const CLIENT_WEBID = "https://id.www.semantyk.com";
+export const POD_PROVIDER_WEBID = "https://pod.danielbakas.com";
+// WebID Documents
+export const APP_WEBID_DOC = "https://storage.inrupt.com/93eee8ab-4002-4bbf-a457-38eab9979cc8/public/profile/card";
