+++ src/frontend/ui/components/atoms/buttons/loginButton.jsx
@@ -1,9 +1,9 @@
 /*
  * –––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
- * # `layout.jsx` | `app`
+ * # `loginButton.jsx` | `buttons`
  * client | Semantyk
  *
- * Created: Nov 30, 2023
+ * Created: Dec 04, 2023
  * Modified: Dec 10, 2023
  *
  * Author(s): Semantyk Team
@@ -13,26 +13,28 @@
  * –––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
  */
 
+"use client";
+
 //* Imports
-import "bootstrap/dist/css/bootstrap.min.css";
-import React from "react";
-//* Local Imports
-import { getMetadata } from "@/backend/logic/seo";
-import StateManager from "@/frontend/logic/state";
+import React, { useEffect, useState } from "react";
+import { LoginButton as SolidLoginButton } from "@inrupt/solid-ui-react";
+import { Button } from "react-bootstrap";
+//* Local imports
+import { getOptions } from "@/backend/logic/auth";
 
 
 //* Main
-export async function generateMetadata() {return await getMetadata();}
-
-export default function RootLayout({ children }) {
+export default function LoginButton() {
+    // Hooks
+    const [options, setOptions] = useState({});
+    // Logic
+    useEffect(() => {
+        getOptions().then(setOptions);
+    }, []);
     // Return
     return (
-        <html lang="en">
-        <body className={`bg-dark text-light`}>
-        <StateManager>
-            {children}
-        </StateManager>
-        </body>
-        </html>
+        <SolidLoginButton {...options}>
+            <Button>Log In</Button>
+        </SolidLoginButton>
     );
 }
