+++ src/frontend/ui/components/molecules/footer/component.jsx
@@ -1,10 +1,10 @@
 /*
  * –––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
- * # `shape.js`
+ * # `component.jsx`
  * client | Semantyk
  *
  * Created: Dec 05, 2023
- * Modified: Dec 05, 2023
+ * Modified: Dec 10, 2023
  *
  * Author(s): Semantyk Team
  * Maintainer(s): Daniel Bakas <https://id.danielbakas.com>
@@ -13,17 +13,25 @@
  * –––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
  */
 
+"use client";
+
+//* Imports
+import Link from "next/link";
 //* Local Imports
-import { SCHEMA } from "../solid/namespaces";
-import { ARCHIVE } from "./namespaces";
+import useKnowledge from "@/frontend/hooks/useKnowledge";
+import FooterLayout from "./layout";
 
 
-//* Main
-export const appShape = [
-    { "ns": SCHEMA, "fragment": "copyrightYear", "type": "number" },
-    { "ns": SCHEMA, "fragment": "description", "type": "string" },
-    { "ns": SCHEMA, "fragment": "keywords", "type": "stringAllNoLocale" },
-    { "ns": SCHEMA, "fragment": "name", "type": "stringNoLocale" },
-    { "ns": SCHEMA, "fragment": "slogan", "type": "string" },
-    { "ns": ARCHIVE, "fragment": "topLevelDomain", "type": "stringNoLocale" }
-];
+export default function Footer() {
+    // Hooks
+    const { name, copyrightYear } = useKnowledge();
+    // Return
+    const link = <Link className="text-secondary" href="/">{name}</Link>;
+    return (
+        <FooterLayout>
+            <small className="text-secondary">
+                Copyright © {link} {copyrightYear}
+            </small>
+        </FooterLayout>
+    );
+}
